crawl/requirements.txt
argcomplete==3.3.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
commitizen==3.27.0
decli==0.6.2
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
prompt-toolkit==3.0.36
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
tomlkit==0.12.5
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.19.2
